[Performance] Optimize GPU kernels and memory layout * Transition dendrite_weights from i16 to i32 in VRAM layout to store full Mass Domain (range up to 2,140,000,000) * Implement Mass/Charge Domain separation: store weights as i32 mass, convert to charge via `w >> 16` in UpdateNeurons kernel * Enforce hard limit of 32 threads per CUDA/AMD block for sort_and_prune kernel to stay within 48 KB Shared Memory budget (12B * 128 slots * 32 threads) * Optimize inject_inputs_kernel dense packing to `words_per_tick_total = (total_num_pixels + 63) / 64 * 2` * Implement branchless burst flag assembly in CUDA: `vram.soma_flags[tid] = (flags & 0xF0) | (burst_count << 1) | final_spike`
[Connectivity] Enhance Python SDK with auto-wiring and zero-copy L7
* Implement Cold Start Auto-Wiring: GenesisIoContract parses baked io.toml to auto-compute C-ABI alignment and L7 fragmentation
* Add Zero-Copy L7 Assembler in GenesisMultiClient: transparently reassembles MTU-chunked UDP payloads with 64-byte alignment
* Enforce 8 MB OS socket buffer (SO_RCVBUF) to prevent UDP overflow from L7 chunk bursts
* Expose SDK Telemetry Translation: get_network_stats() returns avg_weight/max_weight divided by 65536.0 (Charge Domain)

[System] Update specs and examples with new memory calculations
* Update all VRAM calculations in specs to reflect 1166-byte per neuron and i32 dendrite_weights
* Revise Python examples (ant_agent.py, cartpole_exp/agent.py, build_brain.py) to use new GenesisIoContract and GenesisMultiClient
* Correct STDP inertia rank calculation from `abs(weight) >> 11` to `abs(weight) >> 27`
* Document Hebbian Structural Rule: dendrite sprouts only if soma was active (`flags[i] & 0x01 != 0`)

Tested: Updated Python SDK examples, GPU kernel block limits, memory alignment invariants
Refs: specs/01_foundations.md §1.7, specs/07_gpu_runtime.md, docs/Python-SDK/Client_SDK.md §2.2-2.3

+++ genesis-baker/src/bake/layout.rs
@@ -34,7 +34,7 @@ pub struct ShardSoA {
 
     // Транспонированная матрица дендритов (Columnar Layout)
     pub dendrite_targets: Vec<u32>,
-    pub dendrite_weights: Vec<i16>,
+    pub dendrite_weights: Vec<i32>,
     pub dendrite_timers: Vec<u8>, // Refractory timers for synapses
 
     // Аксоны
@@ -133,7 +133,7 @@ pub fn write_state_blob(
     timers: &[u8],
     soma_to_axon: &[u32],
     dendrite_targets: &[u32], // Длина: padded_n * 128
-    dendrite_weights: &[i16], // Длина: padded_n * 128
+    dendrite_weights: &[i32], // Длина: padded_n * 128
     dendrite_timers: &[u8],  // Длина: padded_n * 128
 ) -> std::io::Result<()> {
     let (_, total_size) = calculate_state_blob_size(padded_n);
